firt functions created for labelled data
